Server shell command
This patch defines a shell command to start a server.
